funciones para recibir y hacer transacciones ( hay que ver si estan bien planteadas)

+++ EDAcoin/Definitions.h
@@ -127,7 +127,7 @@ typedef struct {
 
 typedef struct {
 	string publicID;
-	unsigned int amount;
+	double amount;
 }OutputS;
 
 typedef struct {
